requirements.txt
aiogram>=3.10,<4
aiohttp>=3.9,<4
APScheduler>=3.10,<4
motor>=3.5,<4
python-dotenv>=1.0,<2
pymongo>=4.8,<5
httpx>=0.27,<1
